+++ packages/bpk-component-logo/index.js
@@ -1,3 +1,7 @@
-import component from './src/BpkLogo'
+import BpkCloudLogo from './src/BpkCloudLogo'
+import BpkInlineLogo from './src/BpkInlineLogo'
+import BpkStackedLogo from './src/BpkStackedLogo'
+import BpkTianxunLogo from './src/BpkTianxunLogo'
+import BpkTianxunStackedLogo from './src/BpkTianxunStackedLogo'
 
-export default component
+export { BpkCloudLogo, BpkInlineLogo, BpkStackedLogo, BpkTianxunLogo, BpkTianxunStackedLogo }