+++ packages/bpk-component-spinner/src/BpkSpinner-test.js
@@ -1,5 +1,7 @@
 import React from 'react'
 import renderer from 'react/lib/ReactTestRenderer'
+
+import TOKENS from 'bpk-tokens/tokens/base.common'
 import BpkSpinner from './BpkSpinner'
 
 describe('BpkSpinner', () => {
@@ -8,23 +10,13 @@ describe('BpkSpinner', () => {
     expect(tree).toMatchSnapshot()
   })
 
+  it('should render correctly with a "fill" attribute', () => {
+    const tree = renderer.create(<BpkSpinner fill={TOKENS.colorBlue500} />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+
   it('should render correctly with a "alignToButton" attribute', () => {
     const tree = renderer.create(<BpkSpinner alignToButton />).toJSON()
-    expect(tree).toMatchSnapshot()
-  })
-
-  it('should render correctly with a "large" attribute', () => {
-    const tree = renderer.create(<BpkSpinner large />).toJSON()
-    expect(tree).toMatchSnapshot()
-  })
-
-  it('should render correctly with a "large" and a "alignToButton" attribute', () => {
-    const tree = renderer.create(<BpkSpinner large alignToButton />).toJSON()
-    expect(tree).toMatchSnapshot()
-  })
-
-  it('should render correctly with a "extraLarge" attribute', () => {
-    const tree = renderer.create(<BpkSpinner extraLarge />).toJSON()
     expect(tree).toMatchSnapshot()
   })
 })