+++ packages/bpk-component-spinner/stories.js
@@ -1,15 +1,15 @@
 import React from 'react'
 import { storiesOf } from '@kadira/storybook'
 
-import BpkSpinner from './index'
+import { BpkSpinner, BpkLargeSpinner, BpkExtraLargeSpinner } from './index'
 
 storiesOf('bpk-component-spinner', module)
   .add('Small', () => (
     <BpkSpinner />
   ))
   .add('Large', () => (
-    <BpkSpinner large />
+    <BpkLargeSpinner />
   ))
   .add('Extra large', () => (
-    <BpkSpinner extraLarge />
+    <BpkExtraLargeSpinner />
   ))